Stop using pkg/childProcess/spawnFile from scripts.
This is because `spawnFile` depends on `paths.ts` to get logs,
and in `scripts/` files we can't count on having `rdctl` for
the `getPaths` routine.
